OtherProjects/semantic-search-engine/requirements.txt
sentence-transformers
faiss-cpu
streamlit